Model Files
This contains the Netlogo file which is used to run the model as well as
two images.  GallipoliMap is the overlaying image which is used to show
the troops moving around on the peninsula.  GallipoliTopo is an image
for terrain which Netlogo uses to read the different elevations based on
the colors.